Linux build fix: Fix C++17 compatibility: cast int to int64_t for variant types (#124)
The Metadata type uses std::variant with int64_t instead of int.
In C++17, implicit conversion from int to int64_t in variant
assignment is not allowed, requiring explicit static_cast.

Changes:
- Cast all integer literals to int64_t when adding to Metadata
- Update tests to use static_cast<int64_t>() for integer values
- Fix examples to properly cast integers in .add() and .argument() calls
- Ensure compatibility with variant<..., int64_t, ...> type

This fixes compilation errors when building with C++17 standard.
Test results: 91% tests passing (75/82).

+++ examples/test_echo_integration.cc
@@ -93,7 +93,7 @@ int main() {
     std::cout << "\nTest 5: Building Metadata Structures\n";
     auto complex_metadata = mcp::make<mcp::Metadata>()
                                 .add("string_field", "test_string")
-                                .add("int_field", 42)
+                                .add("int_field", static_cast<int64_t>(42))
                                 .add("double_field", 3.14159)
                                 .add("bool_field", true)
                                 .build();
